fix for 256x8 pane
 markruys/arduino-Max72xxPanel#19

+++ src/Max72xxPanel.h
@@ -113,7 +113,7 @@ private:
 
   /* We keep track of the led-status for 8 devices in this array */
   byte *bitmap;
-  byte bitmapSize;
+  uint16_t bitmapSize;
 
   byte hDisplays;
   byte *matrixPosition;
